Removed a bunch of deprecation warnings Cleanup
git-svn-id: svn+ssh://scm.gforge.inria.fr/svn/proactive/trunk@7010 28e8926c-6b08-0410-baaa-805c5e19b8d6

+++ src/Core/org/objectweb/proactive/benchmarks/timit/result/BasicResultWriter.java
@@ -231,7 +231,7 @@ public class BasicResultWriter {
         // Set the name as an attribute
         newTimerElement.setAttribute(new Attribute("name",
                 currentTimer.getName()));
-        double totalTimeValueInMillis = ((double) currentTimer.getTotalTime()) / 1000000d;
+        double totalTimeValueInMillis = (currentTimer.getTotalTime()) / 1000000d;
         // Set the totalTime in millis of the timer as an attribute
         newTimerElement.setAttribute(new Attribute("totalTimeInMillis",
                 "" + totalTimeValueInMillis));
